requirements.txt
pyTelegramBotApi
opencv-python-headless
opencv-python
qrcode
tensorflow
tensorflow-cpu
insightface
onnxruntime
segmentation_models
flask
numpy
Pillow
opendatasets
keras
scipy
deskew
typing
pyyaml
pycocotools
torch
torchvision
cython
pyyaml
pycocotools
pandas
git+https://github.com/qubvel/segmentation_models
git+https://github.com/facebookresearch/detectron2.git
seaborn
